darktable UCS/colorbalance RGB: refactoring
libify the gamut LUT and gamut mapping methods
